refactor: BookHistoryFragment View, Presenter 관심사 분리
