+++ app/src/main/java/woowacourse/movie/fragment/bookhistory/BookHistoryContract.kt
@@ -1,5 +1,6 @@
 package woowacourse.movie.fragment.bookhistory
 
+import woowacourse.movie.database.BookingHistoryRepository
 import woowacourse.movie.movie.MovieBookingSeatInfo
 
 interface BookHistoryContract {
@@ -12,5 +13,6 @@ interface BookHistoryContract {
     interface Presenter {
         fun getData(): List<MovieBookingSeatInfo>
         fun onClickItem(position: Int)
+        fun reloadBookingData(bookingHistory: BookingHistoryRepository)
     }
 }
