+++ app/src/main/java/woowacourse/movie/fragment/bookhistory/BookHistoryPresenter.kt
@@ -1,6 +1,7 @@
 package woowacourse.movie.fragment.bookhistory
 
-import woowacourse.movie.BookHistories
+import woowacourse.movie.database.BookHistories
+import woowacourse.movie.database.BookingHistoryRepository
 import woowacourse.movie.movie.MovieBookingSeatInfo
 
 class BookHistoryPresenter(private val view: BookHistoryContract.View) :
@@ -9,5 +10,12 @@ class BookHistoryPresenter(private val view: BookHistoryContract.View) :
 
     override fun onClickItem(position: Int) {
         view.showDetailPage(position)
+    }
+
+    override fun reloadBookingData(bookingHistory: BookingHistoryRepository) {
+        if (!bookingHistory.isEmpty()) {
+            BookHistories.items.clear()
+            BookHistories.items.addAll(bookingHistory.getAll())
+        }
     }
 }
